+++ src/SMAPI.Web/Framework/Storage/IStorageProvider.cs
@@ -6,13 +6,15 @@ namespace StardewModdingAPI.Web.Framework.Storage;
 internal interface IStorageProvider
 {
     /// <summary>Save a text file to storage.</summary>
+    /// <param name="id">The ID with which to save the result.</param>
     /// <param name="content">The content to upload.</param>
     /// <param name="compress">Whether to gzip the text.</param>
     /// <returns>Returns metadata about the save attempt.</returns>
-    Task<UploadResult> SaveAsync(string content, bool compress = true);
+    Task<UploadResult> SaveAsync(string id, string content, bool compress = true);
 
     /// <summary>Fetch raw text from storage.</summary>
-    /// <param name="id">The storage ID returned by <see cref="SaveAsync"/>.</param>
+    /// <param name="id">The storage ID used to upload the file.</param>
     /// <param name="forceRenew">Whether to reset the file expiry.</param>
-    Task<StoredFileInfo> GetAsync(string id, bool forceRenew);
+    /// <param name="forceDownloadContent">Whether to download the file content even if a fetch URI can be provided.</param>
+    Task<StoredFileInfo> GetAsync(string id, bool forceRenew, bool forceDownloadContent = false);
 }